Fix playlist rowheight
- The tableView:rowHeight: delegate method is supposed to
  return a CGFloat which when building 64-bit is a double
  and not a float. This caused the rows to have zero height.

+++ PlaylistController.m
@@ -229,7 +229,7 @@
     return [_playlist count];
 }
 
-- (float)tableView:(NSTableView*)tableView heightOfRow:(int)rowIndex
+- (CGFloat)tableView:(NSTableView*)tableView heightOfRow:(int)rowIndex
 {
     PlaylistItem* item = [_playlist itemAtIndex:rowIndex];
     int subtitleURLCount = [[item subtitleURLs] count];
